+++ jpro-session/example/src/main/java/module-info.java
@@ -7,9 +7,9 @@ module one.jpro.platform.sessions.example {
     requires javafx.controls;
     requires jpro.webapi;
     requires atlantafx.base;
-    requires one.jpro.platform.sessions;
+    requires one.jpro.platform.session;
     requires org.slf4j;
     requires org.apache.commons.io;
 
-    exports one.jpro.platform.sessions.example;
+    exports one.jpro.platform.session.example;
 }