+++ jpro-session/src/main/java/module-info.java
@@ -1,8 +1,8 @@
-module one.jpro.platform.sessions {
+module one.jpro.platform.session {
     requires javafx.controls;
     requires jpro.webapi;
     requires org.slf4j;
     requires org.apache.commons.io;
 
-    exports one.jpro.platform.sessions;
+    exports one.jpro.platform.session;
 }